requirements.txt
streamlit
transformers
torch
spacy
pdfplumber
python-docx
